refactor(backend): add interface types and refactor foods crud

+++ app/backend/src/controllers/Food.ts
@@ -9,6 +9,7 @@ class FoodController {
     const { id } = req.params
     try {
       const foods = await this.foodService.getAllByTheme(Number(id))
+      if ('error' in foods) return res.status(foods.code).json({ error: foods.error })
       return res.status(200).json(foods)
     } catch (err) {
       return next(err)
@@ -29,6 +30,7 @@ class FoodController {
     const { id } = req.params
     try {
       const updatedFood = await this.foodService.update(Number(id), { price, name, foodType, foodSubType })
+      if ('error' in updatedFood) return res.status(updatedFood.code).json({ error: updatedFood.error })
       return res.status(200).json(updatedFood)
     } catch (err) {
       return next(err)
@@ -48,6 +50,7 @@ class FoodController {
     const { id } = req.params
     try {
       const food = await this.foodService.getFoodById(Number(id))
+      if ('error' in food) return res.status(food.code).json({ error: food.error })
       return res.status(200).json(food)
     } catch (err) {
       return next(err)
@@ -58,6 +61,7 @@ class FoodController {
     const { name, price, foodType, foodSubType } = req.body
     try {
       const newFood = await this.foodService.create({ name, price, foodType, foodSubType })
+      if ('error' in newFood) return res.status(newFood.code).json({ error: newFood.error })
       return res.status(200).json(newFood)
     } catch (err) {
       return next(err)
@@ -69,6 +73,7 @@ class FoodController {
     const { io }: any = req
     try {
       const deletedFood = await this.foodService.delete(Number(id))
+      if ('error' in deletedFood) return res.status(deletedFood.code).json({ error: deletedFood.error })
       deleteFoodFromMenu(io, deletedFood)
       return res.status(200).json(deletedFood)
     } catch (err) {
@@ -81,6 +86,7 @@ class FoodController {
     const { io }: any = req
     try {
       const food = await this.foodService.updateMenu(Number(id))
+      if ('error' in food) return res.status(food.code).json({ error: food.error })
       updateFoodToMenu(io, food)
       io.emit('foodOption-updated', food)
       return res.status(200).json(food)
